# configs/openapi2-run.yaml
...
swagger: '2.0'
info:
  title: Web Service & Discord Bot API Gateway
  description: API Gateway with endpoints for web service and Discord bot on Google Cloud Run backend
  version: 1.0.0
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
paths:
  /discord/interactions:
    post:
      summary: Handle Discord interactions
      description: Webhook endpoint for Discord bot interactions. Routes to Proxy Service which publishes to Pub/Sub topics.
      operationId: discordInteractions
      x-google-backend:
        address: https://proxy-gdu7lmfosq-ew.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: X-Signature-Ed25519
          in: header
          required: true
          type: string
          description: Discord signature for request verification
        - name: X-Signature-Timestamp
          in: header
          required: true
          type: string
          description: Discord timestamp for request verification
        - name: interaction
          in: body
          required: true
          schema:
            type: object
            description: Discord interaction payload
      responses:
        '200':
          description: Successful interaction response
          schema:
            type: object
        '400':
          description: Bad Request - Missing headers or invalid JSON
          schema:
            type: string
        '401':
          description: Unauthorized - Invalid signature
          schema:
            type: string
        '500':
          description: Internal Server Error
          schema:
            type: string
  /web/interactions:
    options:
      summary: CORS preflight for web interactions
      description: Handle CORS preflight requests
      operationId: webInteractionsOptions
      x-google-backend:
        address: https://proxy-gdu7lmfosq-ew.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        '200':
          description: CORS preflight response
    post:
      summary: Handle web interactions
      description: Endpoint for web client interactions. Routes to Proxy Service which publishes to Pub/Sub topics.
      operationId: webInteractions
      x-google-backend:
        address: https://proxy-gdu7lmfosq-ew.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: interaction
          in: body
          required: true
          schema:
            type: object
            description: Web interaction payload
            properties:
              command:
                type: string
                description: Command name
              options:
                type: array
                description: Command options
              token:
                type: string
                description: Interaction token
              application_id:
                type: string
                description: Application ID
      responses:
        '200':
          description: Successful interaction response (immediate)
          schema:
            type: object
        '202':
          description: Command accepted and being processed
          schema:
            type: object
        '400':
          description: Bad Request - Invalid JSON or missing command
          schema:
            type: object
        '500':
          description: Internal Server Error
          schema:
            type: object
        '503':
          description: Service Temporarily Unavailable
          schema:
            type: object
...